% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR-HMC-kde.R
\name{hmc_kde}
\alias{hmc_kde}
\title{Hamiltonian Monte Carlo for Logistic Regression with Optional KDE Priors
Runs a Hamiltonian Monte Carlo (HMC) sampler for logistic regression  using \pkg{nimble} and \pkg{nimbleHMC}. Priors can be standard normal or custom kernel density estimates (KDE) based on posterior samples from a previous model fit.}
\usage{
hmc_kde(
  X,
  y,
  niter = 5000,
  nburnin = 1000,
  nchains = 1,
  inits = NULL,
  KDE = FALSE,
  C = NA
)
}
\arguments{
\item{X}{Numeric matrix of predictors, dimension \eqn{n \times M}, where \eqn{n} is the number of observations and \eqn{M} is the number of predictors.}

\item{y}{Numeric vector of binary responses of length \eqn{n}.}

\item{niter}{Integer. Total number of MCMC iterations per chain (default = 5000).}

\item{nburnin}{Integer. Number of burn-in iterations to discard (default = 1000).}

\item{nchains}{Integer. Number of MCMC chains to run (default = 1).}

\item{inits}{Optional numeric vector of initial values for \code{beta}. If \code{NULL} (default), all coefficients are initialized at zero.}

\item{KDE}{Logical. If \code{FALSE} (default), standard normal priors are used for regression coefficients. If \code{TRUE}, KDE-based priors are constructed from \code{C}.}

\item{C}{Optional numeric matrix of posterior samples from a previous fit, used only when \code{KDE = TRUE}. Each column corresponds to a regression coefficient, and each row is a sample.}
}
\value{
A list containing:
\item{summary}{Summary statistics of posterior samples from all chains.}
\item{samples}{(If \code{summary = FALSE} in \code{\link{runMCMC}}) the raw samples.}
}
\description{
Hamiltonian Monte Carlo for Logistic Regression with Optional KDE Priors
Runs a Hamiltonian Monte Carlo (HMC) sampler for logistic regression  using \pkg{nimble} and \pkg{nimbleHMC}. Priors can be standard normal or custom kernel density estimates (KDE) based on posterior samples from a previous model fit.
}
