% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH.R
\name{log_prior}
\alias{log_prior}
\title{Log-prior density under a multivariate normal distribution
Compute the log-density for a given vector specified from an MVN proposal}
\usage{
log_prior(distro, mu, sigma)
}
\arguments{
\item{distro}{Numeric vector of parameter values.}

\item{mu}{Numeric vector; mean vector of the prior.}

\item{sigma}{Numeric covariance matrix of the prior.}
}
\value{
Numeric scalar: log-prior value
}
\description{
Log-prior density under a multivariate normal distribution
Compute the log-density for a given vector specified from an MVN proposal
}
