% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH.R
\name{logistic_ll}
\alias{logistic_ll}
\title{Log-likelihood of the logistic regression function
Compute the log-likelihood of logistic regression parameters.}
\usage{
logistic_ll(beta, X, y)
}
\arguments{
\item{beta}{Numeric vector of regression coefficients.}

\item{X}{Numeric design matrix.}

\item{y}{Numeric binary response vector (0/1).}
}
\value{
Numeric scalar: log-likelihood value.
}
\description{
Log-likelihood of the logistic regression function
Compute the log-likelihood of logistic regression parameters.
}
