% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR-HMC-kde.R
\name{pr_hmc_kde}
\alias{pr_hmc_kde}
\title{Partitioned Recursive HMC with KDE Priors
Runs a partitioned recursive Hamiltonian Monte Carlo (HMC) logistic regression, where the priors for later partitions are estimated using kdes from the posterior samples of earlier partitions.}
\usage{
pr_hmc_kde(k, X, y, niter = 5000, nburnin = 1000)
}
\arguments{
\item{k}{Integer. Number of partitions to split the dataset into.}

\item{X}{Numeric matrix of predictors (rows = observations, columns = features).}

\item{y}{Numeric vector or single-column matrix of binary outcomes (0 or 1).}

\item{niter}{Integer. Total number of HMC iterations per partition (default = 5000).}

\item{nburnin}{Integer. Number of burn-in iterations to discard per partition (default = 1000).}
}
\value{
A numeric matrix of posterior samples from the last partition.  Columns correspond to regression coefficients (\code{beta}), and rows to iterations.
}
\description{
Partitioned Recursive HMC with KDE Priors
Runs a partitioned recursive Hamiltonian Monte Carlo (HMC) logistic regression, where the priors for later partitions are estimated using kdes from the posterior samples of earlier partitions.
}
