% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP-MH.R
\name{prior_proposal}
\alias{prior_proposal}
\title{Recursive Prior Proposal via Partitioned Metropolis Sampling
Runs a recursive prior proposal process by partitioning the dataset into \code{k} groups and sequentially fitting Metropolis-Hastings models. The first partition is sampled using an MVN prior, while subsequent partitions use random proposals from the posterior of the previous partition}
\usage{
prior_proposal(
  k,
  X,
  y,
  N = 20000,
  M = ncol(X),
  a_target = 0.234,
  b1 = rep(0, M),
  batch_schedule = seq(10, N, by = 10)
)
}
\arguments{
\item{k}{Integer. Number of partitions to split the data into.}

\item{X}{Numeric matrix of predictors (rows = observations, columns = features).}

\item{y}{Numeric vector or matrix of responses (one column).}

\item{N}{Integer. Number of Metropolis samples for the first partition (default = 20000).}

\item{M}{Integer. Number of predictors (default = \code{ncol(X)}).}

\item{a_target}{Numeric. Target acceptance rate for adaptive tuning (default = 0.234).}

\item{b1}{Numeric vector of initial proposal means (default = \code{rep(0, M)}).}

\item{batch_schedule}{Frequency of batching dataset for MH}
}
\value{
A numeric matrix \code{C} containing posterior samples from the last partition's Metropolis-Hastings run.
}
\description{
Recursive Prior Proposal via Partitioned Metropolis Sampling
Runs a recursive prior proposal process by partitioning the dataset into \code{k} groups and sequentially fitting Metropolis-Hastings models. The first partition is sampled using an MVN prior, while subsequent partitions use random proposals from the posterior of the previous partition
}
