% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR-MH.R
\name{prior_recursive}
\alias{prior_recursive}
\title{Recursive prior estimation with adaptive Metropolis sampling
Performs a recursive Bayesian update where data are split into \code{k} partitions. The posterior from one partition is used to form a kernel density estimate (KDE) prior for the next partition. Sampling for each stage is performed using the \code{\link{metropolis}} function.}
\usage{
prior_recursive(
  k,
  X,
  y,
  N = 20000,
  M = ncol(X),
  a_target = 0.234,
  b1 = rep(0, M),
  batch_schedule = seq(10, N, by = 10)
)
}
\arguments{
\item{k}{Integer; number of data partitions.}

\item{X}{Numeric design matrix of predictors.}

\item{y}{Numeric response vector (binary outcomes 0/1).}

\item{N}{Integer; number of MCMC iterations per partition (default \code{20000}).}

\item{a_target}{Numeric; target acceptance rate for Metropolis sampler (default \code{0.234}).}

\item{b1}{Numeric vector; initial parameter values for the first partition (default \code{rep(0, M)}).}

\item{batch_schedule}{Frequency of batching dataset for MH}
}
\value{
Numeric matrix of posterior samples from the final stage after all recursive updates.
}
\description{
Recursive prior estimation with adaptive Metropolis sampling
Performs a recursive Bayesian update where data are split into \code{k} partitions. The posterior from one partition is used to form a kernel density estimate (KDE) prior for the next partition. Sampling for each stage is performed using the \code{\link{metropolis}} function.
}
\details{
The data are split row-wise into \code{k} roughly equal groups. For the first group, the Metropolis sampler is run with a MVN prior. For subsequent groups, KDEs are fit to the posterior samples from the previous stage, producing univariate priors for each parameter. The prior mean vector is set to the mode of each KDE for the starting values.
}
